+++ sdk-java/src/main/java/ly/count/sdk/java/Config.java
@@ -13,21 +13,23 @@ import java.util.Map;
 import java.util.Set;
 
 import ly.count.sdk.java.internal.*;
+import ly.count.sdk.java.internal.Module;
 
 /**
  * Countly configuration object.
  */
 public class Config {
-    private static final Log.Module L = Log.module("ConfigCore");
+
     /**
      * Logging level for {@link Log} module
      */
     public enum LoggingLevel {
-        DEBUG(0),
-        INFO(1),
-        WARN(2),
-        ERROR(3),
-        OFF(4);
+        VERBOSE(0),
+        DEBUG(1),
+        INFO(2),
+        WARN(3),
+        ERROR(4),
+        OFF(5);
 
         private final int level;
 
@@ -160,20 +162,20 @@ public class Config {
                 stream.close();
                 return bytes.toByteArray();
             } catch (IOException e) {
-                L.wtf("Cannot serialize config", e);
+                System.out.print("[ConfigCore] Cannot serialize config" + e.toString());
             } finally {
                 if (stream != null) {
                     try {
                         stream.close();
                     } catch (IOException e) {
-                        L.wtf("Cannot happen", e);
+                        System.out.print("[ConfigCore] Cannot happen" + e.toString());
                     }
                 }
                 if (bytes != null) {
                     try {
                         bytes.close();
                     } catch (IOException e) {
-                        L.wtf("Cannot happen", e);
+                        System.out.print("[ConfigCore] Cannot happen" + e.toString());
                     }
                 }
             }
@@ -195,20 +197,20 @@ public class Config {
 
                 return true;
             } catch (IOException | ClassNotFoundException e) {
-                L.wtf("Cannot deserialize config", e);
+                System.out.print("[ConfigCore] Cannot deserialize config" + e.toString());
             } finally {
                 if (stream != null) {
                     try {
                         stream.close();
                     } catch (IOException e) {
-                        L.wtf("Cannot happen", e);
+                        System.out.print("[ConfigCore] Cannot happen" + e.toString());
                     }
                 }
                 if (bytes != null) {
                     try {
                         bytes.close();
                     } catch (IOException e) {
-                        L.wtf("Cannot happen", e);
+                        System.out.print("[ConfigCore] Cannot happen" + e.toString());
                     }
                 }
             }
@@ -256,6 +258,12 @@ public class Config {
      * Logging level
      */
     protected LoggingLevel loggingLevel = LoggingLevel.OFF;
+
+    /**
+     * Log listener
+     */
+    protected LogCallback logListener = null;
+
 
     /**
      * Countly SDK name to be sent in HTTP requests
@@ -534,11 +542,11 @@ public class Config {
      */
     public Config enableFeatures(Config.Feature... features) {
         if (features == null) {
-            L.wtf("Features array cannot be null");
+            System.out.print("[ConfigCore] Features array cannot be null");
         } else {
             for (Config.Feature f : features) {
                 if (f == null) {
-                    L.wtf("Feature cannot be null");
+                    System.out.print("[ConfigCore] Feature cannot be null");
                 } else {
                     this.features = this.features | f.getIndex();
                 }
@@ -555,16 +563,28 @@ public class Config {
      */
     public Config disableFeatures(Config.Feature... features) {
         if (features == null) {
-            L.wtf("Features array cannot be null");
+            System.out.print("[ConfigCore] Features array cannot be null");
         } else {
             for (Config.Feature f : features) {
                 if (f == null) {
-                    L.wtf("Feature cannot be null");
+                    System.out.print("[ConfigCore] Feature cannot be null");
                 } else {
                     this.features = this.features & ~f.getIndex();
                 }
             }
         }
+        return this;
+    }
+
+    /**
+     * Add a log callback that will duplicate all logs done by the SDK.
+     * For each message you will receive the message string and it's targeted log level.
+     *
+     * @param logCallback
+     * @return Returns the same config object for convenient linking
+     */
+    public Config setLogListener(LogCallback logCallback) {
+        this.logListener = logCallback;
         return this;
     }
 
@@ -580,7 +600,7 @@ public class Config {
         if (features != null && features.length > 0) {
             for (int i = 0; i < features.length; i++) {
                 if (features[i] == null) {
-                    L.wtf(i + "-th feature is null in setFeatures");
+                    System.out.print(i + "-th feature is null in setFeatures");
                 } else {
                     this.features = this.features | features[i].index;
                 }
@@ -600,7 +620,7 @@ public class Config {
      */
     public Config setDeviceIdStrategy(DeviceIdStrategy strategy, String customDeviceId) {
         if (strategy == null) {
-            L.wtf("DeviceIdStrategy cannot be null");
+            System.out.print("[ConfigCore] DeviceIdStrategy cannot be null");
         } else {
             if (strategy == DeviceIdStrategy.CUSTOM_ID) {
                 return setCustomDeviceId(customDeviceId);
@@ -628,7 +648,7 @@ public class Config {
      */
     public Config setCustomDeviceId(String customDeviceId) {
         if (Utils.isEmptyOrNull(customDeviceId)) {
-            L.wtf("DeviceIdStrategy.CUSTOM_ID strategy cannot be used without device id specified");
+            System.out.print("[ConfigCore] DeviceIdStrategy.CUSTOM_ID strategy cannot be used without device id specified");
         } else {
             this.customDeviceId = customDeviceId;
             this.deviceIdStrategy = DeviceIdStrategy.CUSTOM_ID.index;
@@ -697,7 +717,7 @@ public class Config {
      */
     public Config enableParameterTamperingProtection(String salt) {
         if (Utils.isEmptyOrNull(salt)) {
-            L.wtf("Salt cannot be empty in enableParameterTamperingProtection");
+            System.out.print("[ConfigCore] Salt cannot be empty in enableParameterTamperingProtection");
         } else {
             this.salt = salt;
         }
@@ -712,7 +732,7 @@ public class Config {
      */
     public Config setLoggingTag(String loggingTag) {
         if (loggingTag == null || loggingTag.equals("")) {
-            L.wtf("Logging tag cannot be empty");
+            System.out.print("[ConfigCore] Logging tag cannot be empty");
         } else {
             this.loggingTag = loggingTag;
         }
@@ -727,7 +747,7 @@ public class Config {
      */
     public Config setLoggingLevel(LoggingLevel loggingLevel) {
         if (loggingLevel == null) {
-            L.wtf("Logging level cannot be null");
+            System.out.print("[ConfigCore] Logging level cannot be null");
         } else {
             this.loggingLevel = loggingLevel;
         }
@@ -746,7 +766,7 @@ public class Config {
      *
      * @return {@code this} instance for method chaining
      */
-    public Config enableTestMode() {
+    protected Config enableTestMode() {
         this.testMode = true;
         this.loggingLevel = this.loggingLevel == LoggingLevel.OFF ? LoggingLevel.INFO : this.loggingLevel;
         return this;
@@ -758,7 +778,7 @@ public class Config {
      *
      * @return {@code this} instance for method chaining
      */
-    public Config disableTestMode() {
+    protected Config disableTestMode() {
         this.testMode = false;
         return this;
     }
@@ -774,7 +794,7 @@ public class Config {
      */
     public Config setSendUpdateEachSeconds(int sendUpdateEachSeconds) {
         if (sendUpdateEachSeconds < 0) {
-            L.wtf("sendUpdateEachSeconds cannot be negative");
+            System.out.print("[ConfigCore] sendUpdateEachSeconds cannot be negative");
         } else {
             this.sendUpdateEachSeconds = sendUpdateEachSeconds;
         }
@@ -791,7 +811,7 @@ public class Config {
      */
     public Config setEventsBufferSize(int eventsBufferSize) {
         if (eventsBufferSize < 0) {
-            L.wtf("eventsBufferSize cannot be negative");
+            System.out.print("[ConfigCore] eventsBufferSize cannot be negative");
         } else {
             this.eventsBufferSize = eventsBufferSize;
         }
@@ -820,7 +840,7 @@ public class Config {
      */
     public Config setSessionCooldownPeriod(int sessionCooldownPeriod) {
         if (sessionCooldownPeriod < 0) {
-            L.wtf("sessionCooldownPeriod cannot be negative");
+            System.out.print("[ConfigCore] sessionCooldownPeriod cannot be negative");
         } else {
             this.sessionCooldownPeriod = sessionCooldownPeriod;
         }
@@ -835,7 +855,7 @@ public class Config {
      */
     public Config setSdkName(String sdkName) {
         if (Utils.isEmptyOrNull(sdkName)) {
-            L.wtf("sdkName cannot be empty");
+            System.out.print("[ConfigCore] sdkName cannot be empty");
         } else {
             this.sdkName = sdkName;
         }
@@ -850,7 +870,7 @@ public class Config {
      */
     public Config setSdkVersion(String sdkVersion) {
         if (Utils.isEmptyOrNull(sdkVersion)) {
-            L.wtf("sdkVersion cannot be empty");
+            System.out.print("[ConfigCore] sdkVersion cannot be empty");
         } else {
             this.sdkVersion = sdkVersion;
         }
@@ -865,7 +885,7 @@ public class Config {
      */
     public Config setApplicationName(String name) {
         if (Utils.isEmptyOrNull(name)) {
-            L.wtf("name cannot be empty");
+            System.out.print("[ConfigCore] name cannot be empty");
         } else {
             this.applicationName = name;
         }
@@ -880,7 +900,7 @@ public class Config {
      */
     public Config setApplicationVersion(String version) {
         if (Utils.isEmptyOrNull(version)) {
-            L.wtf("version cannot be empty");
+            System.out.print("[ConfigCore] version cannot be empty");
         } else {
             this.applicationVersion = version;
         }
@@ -895,7 +915,7 @@ public class Config {
      */
     public Config setNetworkConnectTimeout(int seconds) {
         if (seconds <= 0 || seconds > 300) {
-            L.wtf("Connection timeout must be between 0 and 300");
+            System.out.print("[ConfigCore] Connection timeout must be between 0 and 300");
         } else {
             networkConnectionTimeout = seconds;
         }
@@ -910,7 +930,7 @@ public class Config {
      */
     public Config setNetworkReadTimeout(int seconds) {
         if (seconds <= 0 || seconds > 300) {
-            L.wtf("Read timeout must be between 0 and 300");
+            System.out.print("[ConfigCore] Read timeout must be between 0 and 300");
         } else {
             networkReadTimeout = seconds;
         }
@@ -926,7 +946,7 @@ public class Config {
      */
     public Config setNetworkRequestCooldown(int milliseconds) {
         if (milliseconds < 0 || milliseconds > 30000) {
-            L.wtf("Request cooldown must be between 0 and 30000");
+            System.out.print("[ConfigCore] Request cooldown must be between 0 and 30000");
         } else {
             networkRequestCooldown = milliseconds;
         }
@@ -942,7 +962,7 @@ public class Config {
      */
     public Config setNetworkImportantRequestCooldown(int milliseconds) {
         if (milliseconds < 0 || milliseconds > 30) {
-            L.wtf("Important request cooldown must be between 0 and 30");
+            System.out.print("[ConfigCore] Important request cooldown must be between 0 and 30");
         } else {
             networkImportantRequestCooldown = milliseconds;
         }
@@ -970,7 +990,7 @@ public class Config {
      */
     public Config addPublicKeyPin(String pemEncodedPublicKey) {
         if (Utils.isEmptyOrNull(pemEncodedPublicKey)) {
-            L.wtf("pemEncodedPublicKey cannot be empty");
+            System.out.print("[ConfigCore] pemEncodedPublicKey cannot be empty");
         } else {
             if (publicKeyPins == null) {
                 publicKeyPins = new HashSet<>();
@@ -1002,7 +1022,7 @@ public class Config {
      */
     public Config addCertificatePin(String pemEncodedCertificate) {
         if (Utils.isEmptyOrNull(pemEncodedCertificate)) {
-            L.wtf("pemEncodedCertificate cannot be empty");
+            System.out.print("[ConfigCore] pemEncodedCertificate cannot be empty");
         } else {
             if (certificatePins == null) {
                 certificatePins = new HashSet<>();
@@ -1028,7 +1048,7 @@ public class Config {
      */
     public Config setCrashReportingANRCheckingPeriod(int periodInSeconds) {
         if (periodInSeconds < 0) {
-            L.wtf("ANR timeout less than zero doesn't make sense");
+            System.out.print("[ConfigCore] ANR timeout less than zero doesn't make sense");
         } else {
             this.crashReportingANRCheckingPeriod = periodInSeconds;
         }
@@ -1054,7 +1074,7 @@ public class Config {
      */
     public Config setCrashProcessorClass(Class<? extends CrashProcessor> crashProcessorClass) {
         if (crashProcessorClass == null) {
-            L.wtf("crashProcessorClass cannot be null");
+            System.out.print("[ConfigCore] crashProcessorClass cannot be null");
         } else {
             this.crashProcessorClass = crashProcessorClass.getName();
         }
@@ -1070,7 +1090,7 @@ public class Config {
      */
     protected Config overrideModule(Integer feature, Class<? extends Module> cls) {
         if (feature == null || cls == null) {
-            L.wtf("Feature & class cannot be null");
+            System.out.print("[ConfigCore] Feature & class cannot be null");
         } else {
             if (moduleOverrides == null) {
                 moduleOverrides = new HashMap<>();
@@ -1311,10 +1331,18 @@ public class Config {
     }
 
     /**
+     * Getter for {@link #logListener}
+     * @return {@link #logListener} value
+     */
+    public LogCallback getLogListener() {
+        return logListener;
+    }
+
+    /**
      * Getter for {@link #testMode}
      * @return {@link #testMode} value
      */
-    public boolean isTestModeEnabled() {
+    protected boolean isTestModeEnabled() {
         return testMode;
     }